Fix defancifying of titles
Remove symbols from title parts before searching for them in the site

In the case of wikipedia, the title would be split into eg:
['Flubber', 'Wikipedia, the free encyclopedia'], and then compared to
'en.wikipedia.org'

But 'wikipedia,' (note the comma) is not in 'en.wikipedia.org'.  So now
punctuation is removed from the title when comparing to the site.